Fixed distance in GetScore being an int instead of float

+++ SnakeGame.cs
@@ -266,7 +266,7 @@ namespace Snake
         /// <returns>The current score of the game</returns>
         public float GetScore()
         {
-            int dist = (int)Math.Sqrt(Math.Pow(headLocation[0] - appleLocation[0], 2) + Math.Pow(headLocation[1] - appleLocation[1], 2));
+            float dist = (float)Math.Sqrt(Math.Pow(headLocation[0] - appleLocation[0], 2) + Math.Pow(headLocation[1] - appleLocation[1], 2));
             if (dist != 0)
             {
                 dist = 1 / dist;
